Use exoplayer composable for internal and external video dialogs
