+++ ui/src/main/java/com/neaniesoft/vermilion/ui/videos/external/ExternalVideoDialog.kt
@@ -2,18 +2,16 @@ package com.neaniesoft.vermilion.ui.videos.external
 
 import android.net.Uri
 import android.util.Log
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Modifier
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.neaniesoft.vermilion.ui.images.rememberZoomableState
-import com.neaniesoft.vermilion.ui.videos.Video
-import com.neaniesoft.vermilion.ui.videos.VideoPlayer
 import com.neaniesoft.vermilion.ui.videos.ZoomableDialog
+import com.neaniesoft.vermilion.ui.videos.exoplayer.ExoPlayerWithControls
+import com.neaniesoft.vermilion.ui.videos.exoplayer.rememberExoPlayerState
 import com.neaniesoft.vermilion.ui.videos.rememberVideoPlayerState
 
 @ExperimentalMaterialApi
@@ -26,6 +24,7 @@ fun ExternalVideoDialog(
     val uiState by viewModel.uiState.collectAsState()
     val zoomableState = rememberZoomableState(maxScale = 6f)
     val videoPlayerState = rememberVideoPlayerState()
+    val exoPlayerState = rememberExoPlayerState()
 
     LaunchedEffect(key1 = unresolvedUri) {
         viewModel.onResolveExternalUri(unresolvedUri)
@@ -41,11 +40,8 @@ fun ExternalVideoDialog(
                 ) // TODO Handle this better
             }
             is ExternalVideoDialogState.PlayUriState -> {
-                VideoPlayer(
-                    state = videoPlayerState,
-                    video = Video.UriVideo(currentState.uri),
-                    modifier = Modifier.fillMaxSize()
-                )
+                exoPlayerState.mediaUri = currentState.uri
+                ExoPlayerWithControls(exoPlayerState)
             }
         }
     }
